+++ src/index.js
@@ -1,6 +1,3 @@
-import mathSign from './polyfills/math-sign'
-import raf from './polyfills/raf'
-
 import deepAssign from './utils/deep-assign'
 import each from './utils/each'
 import getNode from './utils/get-node'
@@ -9,4 +6,4 @@ import isNodeList from './utils/is-node-list'
 import isNode from './utils/is-node'
 import isObject from './utils/is-object'
 
-export { mathSign, raf, deepAssign, each, getNode, getNodes, isNodeList, isNode, isObject }
+export { deepAssign, each, getNode, getNodes, isNodeList, isNode, isObject }