+++ src/polyfills/math-sign.js
@@ -1,2 +1,0 @@
-export const polyfill = x => (x > 0) - (x < 0) || +x
-export default Math.sign || polyfill